pipeline/pyproject.toml
[project]
name = "guitar-tone-shootout-pipeline"
version = "0.1.0"
description = "Audio/video processing pipeline for guitar amp/cab comparisons"
readme = "../README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
keywords = ["guitar", "amp", "nam", "neural-amp-modeler", "audio", "video"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Multimedia :: Video",
]

dependencies = [
    "pedalboard>=0.9.0",
    "neural-amp-modeler>=0.12.0",
    "numba>=0.59.0",  # Force compatible version for Python 3.12
    "numpy>=1.26.0",
    "pillow>=10.0.0",
    "jinja2>=3.1.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "playwright>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.4.0",
]

[project.scripts]
shootout = "guitar_tone_shootout.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/guitar_tone_shootout"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented code)
    "PL",     # pylint
    "RUF",    # ruff-specific
]
ignore = [
    "PLR0913",  # too many arguments
    "PLR2004",  # magic value comparison
    "PLC0415",  # import not at top-level (intentional for lazy loading heavy deps)
]

[tool.ruff.lint.isort]
known-first-party = ["guitar_tone_shootout"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = false
disallow_untyped_defs = true
check_untyped_defs = true
files = ["src", "tests"]

[[tool.mypy.overrides]]
module = ["pedalboard", "pedalboard.*", "nam.*", "playwright.*", "torch.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests.*", "conftest"]
disallow_untyped_decorators = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "unit: Fast unit tests with mocked dependencies",
    "integration: Integration tests requiring FFmpeg and real files",
    "e2e: End-to-end tests requiring all dependencies",
    "slow: Tests that take more than 5 seconds",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src/guitar_tone_shootout"]
branch = true
